+++ examples/bare/android/app/build.gradle
@@ -73,9 +73,9 @@ android {
 
     compileSdkVersion rootProject.ext.compileSdkVersion
 
-    namespace "com.amabareexample"
+    namespace "com.example.rnama"
     defaultConfig {
-        applicationId "com.amabareexample"
+        applicationId "com.example.rnama"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
