+++ examples/bare/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">AmaBareExample</string>
+    <string name="app_name">AMA Example</string>
 </resources>
