+++ examples/bare/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "AmaExample",
-  "displayName": "Ama Example"
+  "name": "AMA Example",
+  "displayName": "AMA Example"
 }
