+++ examples/bare/ios/AMAExample.xcodeproj/xcshareddata/xcschemes/AMAExample.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "AmaBareExample.app"
-               BlueprintName = "AmaBareExample"
-               ReferencedContainer = "container:AmaBareExample.xcodeproj">
+               BuildableName = "AMAExample.app"
+               BlueprintName = "AMAExample"
+               ReferencedContainer = "container:AMAExample.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "AmaBareExampleTests.xctest"
-               BlueprintName = "AmaBareExampleTests"
-               ReferencedContainer = "container:AmaBareExample.xcodeproj">
+               BuildableName = "AMAExampleTests.xctest"
+               BlueprintName = "AMAExampleTests"
+               ReferencedContainer = "container:AMAExample.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "AmaBareExample.app"
-            BlueprintName = "AmaBareExample"
-            ReferencedContainer = "container:AmaBareExample.xcodeproj">
+            BuildableName = "AMAExample.app"
+            BlueprintName = "AMAExample"
+            ReferencedContainer = "container:AMAExample.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "AmaBareExample.app"
-            BlueprintName = "AmaBareExample"
-            ReferencedContainer = "container:AmaBareExample.xcodeproj">
+            BuildableName = "AMAExample.app"
+            BlueprintName = "AMAExample"
+            ReferencedContainer = "container:AMAExample.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
