+++ examples/bare/ios/AMAExample/AppDelegate.h
