+++ examples/bare/ios/AMAExample/Images.xcassets/AppIcon.appiconset/Contents.json
