+++ examples/bare/ios/AMAExample/Images.xcassets/Contents.json
