+++ examples/bare/ios/AMAExample/main.m
