+++ examples/bare/ios/AMAExampleTests/AMAExampleTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface AmaBareExampleTests : XCTestCase
+@interface AMAExampleTests : XCTestCase
 
 @end
 
-@implementation AmaBareExampleTests
+@implementation AMAExampleTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL (^)(UIView *view))test
 {
