+++ examples/bare/ios/AMAExampleTests/Info.plist
