+++ examples/bare/ios/Podfile
@@ -25,7 +25,7 @@ if linkage != nil
   use_frameworks! :linkage => linkage.to_sym
 end
 
-target 'AmaBareExample' do
+target 'AMAExample' do
   config = use_native_modules!
 
   # Flags change depending on the env values.
@@ -45,7 +45,7 @@ target 'AmaBareExample' do
     :app_path => "#{Pod::Config.instance.installation_root}/.."
   )
 
-  target 'AmaBareExampleTests' do
+  target 'AMAExampleTests' do
     inherit! :complete
     # Pods for testing
   end
