+++ examples/bare/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "AmaBareExample",
+  "name": "amaexample",
   "version": "0.0.1",
   "private": true,
   "scripts": {
